+++ packages/expo-blur/src/BlurTargetView.tsx
@@ -1,0 +1,7 @@
+import { View } from 'react-native';
+
+import { BlurTargetViewProps } from './BlurView.types';
+
+export default function BlurTargetView(props: BlurTargetViewProps) {
+  return <View {...props} />;
+}