+++ packages/expo-blur/src/BlurView.tsx
@@ -2,16 +2,24 @@
 
 'use client';
 
-import { requireNativeViewManager } from 'expo-modules-core';
 import React from 'react';
-import { View, StyleSheet } from 'react-native';
+import { View, StyleSheet, findNodeHandle, Platform } from 'react-native';
 
 import { BlurViewProps } from './BlurView.types';
+import { NativeBlurView } from './NativeBlurModule';
 
-const NativeBlurView = requireNativeViewManager('ExpoBlurView');
+type BlurViewState = {
+  blurTargetId?: number | null;
+};
 
 // TODO: Class components are not supported with React Server Components.
-export default class BlurView extends React.Component<BlurViewProps> {
+export default class BlurView extends React.Component<BlurViewProps, BlurViewState> {
+  constructor(props: BlurViewProps) {
+    super(props);
+    this.state = {
+      blurTargetId: undefined,
+    };
+  }
   blurViewRef? = React.createRef<typeof NativeBlurView>();
 
   /**
@@ -22,6 +30,35 @@ export default class BlurView extends React.Component<BlurViewProps> {
   getAnimatableRef() {
     return this.blurViewRef?.current;
   }
+
+  componentDidMount(): void {
+    this._updateBlurTargetId();
+    this._maybeWarnAboutBlurMethod();
+  }
+
+  componentDidUpdate(prevProps: Readonly<BlurViewProps>): void {
+    if (prevProps.blurTarget?.current !== this.props.blurTarget?.current) {
+      this._updateBlurTargetId();
+    }
+  }
+
+  _maybeWarnAboutBlurMethod(): void {
+    const blurMethod = this.props.experimentalBlurMethod ?? 'none';
+    if (Platform.OS === 'android' && blurMethod === 'dimezisBlurView' && !this.props.blurTarget) {
+      // The fallback happens on the native side
+      console.warn(
+        `You have selected the "${blurMethod}" blur method, but the \`blurTarget\` prop has not been configured. The blur view will fallback to "none" blur method to avoid errors. You can learn more about the new BlurView API at: https://docs.expo.dev/versions/latest/sdk/blur-view/`
+      );
+    }
+  }
+
+  _updateBlurTargetId = () => {
+    const blurTarget = this.props.blurTarget?.current;
+    const blurTargetId = blurTarget ? findNodeHandle(blurTarget) : undefined;
+    this.setState(() => ({
+      blurTargetId,
+    }));
+  };
 
   render() {
     const {
@@ -36,6 +73,7 @@ export default class BlurView extends React.Component<BlurViewProps> {
     return (
       <View {...props} style={[styles.container, style]}>
         <NativeBlurView
+          blurTargetId={this.state.blurTargetId}
           ref={this.blurViewRef}
           tint={tint}
           intensity={intensity}